Restructuring projects (the .NET Core migration left something wrong, I couldn't find it, so I decided to recreated the projects from scratch)
