--- NHiLo.sln
@@ -1,98 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30709.132
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0743066B-0FEC-48D2-9A77-804A51ECC55F}"
-	ProjectSection(SolutionItems) = preProject
-		tests\Local.testsettings = tests\Local.testsettings
-		NHiLo.vsmdi = NHiLo.vsmdi
-		nuget\NuGet.exe = nuget\NuGet.exe
-		readme.md = readme.md
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E2195FEA-9849-45FC-BC6A-B4AD6AD3D392}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{10A18B4F-0D57-48AF-927B-F7D2DBDA7512}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHiLo", "src\NHiLo\NHiLo.csproj", "{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHiLo.Example", "tests\NHiLo.Example\NHiLo.Example.csproj", "{3C9F4024-9FE2-4332-B293-2122136F2902}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHiLo.SqlSequence.QuickTestExample", "tests\NHiLo.SqlSequence.QuickTestExample\NHiLo.SqlSequence.QuickTestExample.csproj", "{1EB09A27-971E-46EB-9A8D-A929D94B15BB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHiLo.Tests", "tests\NHiLo.Tests\NHiLo.Tests.csproj", "{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Debug|x86.Build.0 = Debug|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|x86.ActiveCfg = Release|Any CPU
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E}.Release|x86.Build.0 = Release|Any CPU
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Debug|x86.ActiveCfg = Debug|x86
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Debug|x86.Build.0 = Debug|x86
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Release|Mixed Platforms.Build.0 = Release|x86
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Release|x86.ActiveCfg = Release|x86
-		{3C9F4024-9FE2-4332-B293-2122136F2902}.Release|x86.Build.0 = Release|x86
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Debug|x86.Build.0 = Debug|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|x86.ActiveCfg = Release|Any CPU
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB}.Release|x86.Build.0 = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Debug|x86.Build.0 = Debug|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|Any CPU.Build.0 = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|x86.ActiveCfg = Release|Any CPU
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{3FB860C7-CC27-4865-ACD4-0C8E2903AE9E} = {E2195FEA-9849-45FC-BC6A-B4AD6AD3D392}
-		{3C9F4024-9FE2-4332-B293-2122136F2902} = {10A18B4F-0D57-48AF-927B-F7D2DBDA7512}
-		{1EB09A27-971E-46EB-9A8D-A929D94B15BB} = {10A18B4F-0D57-48AF-927B-F7D2DBDA7512}
-		{54C5832B-4A8D-4294-9E84-FD6AD9D8CE29} = {10A18B4F-0D57-48AF-927B-F7D2DBDA7512}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1D1477E3-5EF6-4181-A483-9A6ABAC4A7DB}
-	EndGlobalSection
-	GlobalSection(TestCaseManagementSettings) = postSolution
-		CategoryFile = NHiLo.vsmdi
-	EndGlobalSection
-EndGlobal
