+++ src/NHiLo.Tests/Integration/Hilo/HiLoGeneratorFactoryTest.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Text;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using NHiLo.HiLo;
+using Microsoft.Extensions.Configuration;
+
+namespace NHiLo.Tests.Integration.Hilo
+{
+    public class HiLoGeneratorFactoryTest
+    {
+        public class GetKeyGenerator
+        {
+            /*
+            [Fact]
+            public void ShouldReturnAValidKey()
+            {
+                // Arrange
+                var builder = new ConfigurationBuilder()
+                    .SetBasePath(Environment.CurrentDirectory)
+                    .AddJsonFile("appsettings.json", optional: true, reloadOnChange: false);
+                var factory = new HiLoGeneratorFactory(builder.Build());
+                // Act
+                var generator = factory.GetKeyGenerator("dummy5");
+                var key1 = generator.GetKey();
+                var key2 = generator.GetKey();
+                // Assert
+                Assert.True(key1 > 0);
+                Assert.True(key2 > 0);
+            }
+            */
+        }
+    }
+}
