+++ src/NHilo.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30804.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHilo", "NHilo\NHilo.csproj", "{49EDF44C-547F-45EC-8C1C-796AD9046FAB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NHiLo.Tests", "NHiLo.Tests\NHiLo.Tests.csproj", "{FCC18C67-323F-4EF6-9C99-522BDFA4CCE8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{49EDF44C-547F-45EC-8C1C-796AD9046FAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{49EDF44C-547F-45EC-8C1C-796AD9046FAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{49EDF44C-547F-45EC-8C1C-796AD9046FAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{49EDF44C-547F-45EC-8C1C-796AD9046FAB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FCC18C67-323F-4EF6-9C99-522BDFA4CCE8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FCC18C67-323F-4EF6-9C99-522BDFA4CCE8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FCC18C67-323F-4EF6-9C99-522BDFA4CCE8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FCC18C67-323F-4EF6-9C99-522BDFA4CCE8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2B4529A7-B710-429D-AB7F-11F4D1D5BDEA}
+	EndGlobalSection
+EndGlobal
