--- tests/NHiLo.Tests/App.config
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <configSections>
-    <section name="nhilo" type="NHiLo.Common.KeyGeneratorConfig, NHiLo, Version=1.1.0.0, Culture=neutral, PublicKeyToken=null"/>
-  </configSections>
-  <connectionStrings>
-    <!--<add name="NHiLo" connectionString="Server=.\SQLEXPRESS;Database=TesteCoda;Trusted_Connection=True;" providerName="System.Data.SqlClient"/>-->
-    <add name="NHiLo" connectionString="Data Source=|DataDirectory|\Database1.sdf;Persist Security Info=False;" providerName="System.Data.SqlServerCe.4.0"/>
-  </connectionStrings>
-  <nhilo>
-    <hiloKeyGenerator>
-      
-    </hiloKeyGenerator>
-  </nhilo>
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/></startup></configuration>
