--- tests/NHiLo.Tests/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "ConnectionStrings": {
-    "NHiLo": {
-      "ConnectionString": "Data Source=|DataDirectory|\\Database1.sdf;Persist Security Info=False;",
-      "ProviderName": "System.Data.SqlServerCe.4.0"
-    }
-  },
-  "nhilo": {
-    "hiloKeyGenerator": {
-
-    }
-  }
-}
