--- tests/NHiLo.Tests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Moq" version="4.0.10827" targetFramework="net40" />
-</packages>
